NOTE: Intermediate commit before file renaming.
CHG: Global Shader variables such as pixel size, transformation matrices and etc. added for all rendering stages.
ADD: ZEWindow is improved and moved to graphics module. Windows are now managed by graphics module.
IMP: D3D 10 Module ported to D3D 11.
CHG: Debug drawing is disabled for initial renderer tryouts.
ADD: Variable number of buffering of D3D resource data is implemented for fast read and write modifications.
IMP: Graphics Resource statistics(Texture count, size) implemented.
CHG: Anisotropic filtering and anti aliasing related methods are moved to graphics device.
ADD: All kinds of shader compilation support is added how ever their pipeline stages are not implemented yet.
IMP: ZEView is implemented as a parent class of camera and other view types such as shadows.
IMP: Property changed flag is added to light entities for graphics performance reasons.

svn path=/branches/v0.6.1-NewGraphics/; revision=1522

+++ Source/ZEFoundation/ZERandom.cpp
@@ -119,7 +119,7 @@ static int Mix(int a, int b, int c)
 
 void ZERandom::Reset()
 {
-	srand(Mix(getpid(), time(NULL), clock()));
+	srand(Mix(getpid(), (int)time(NULL), (int)clock()));
 }
 
 class ZERandomInitializer
